Automatically backed up by Learn

+++ index.js
@@ -31,5 +31,5 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
-return parseInt(n)
+return parseFloat(n)
 }
